Update style.css
I have added a css code snippet from line 343 and onwards. The code is for floating chatbot that opens and closes, have a send button that enters the users queries by pressing "Enter" key. As of now, the code that I have added is running on local model (gemma3:4b) I was not able to integrate Gemini or Deepseek model through API calling( if you are able to do that please do that)